deploy_temp/requirements.txt
fastapi
uvicorn
python-jose
python-multipart
pandas
numpy
ta
requests
websockets
python-dotenv
ccxt
yfinance
